Only download proton binutils patches if they are not found in build dir. Since we're targeting a specific commit id, the patches are static so no reason to refresh them. In case of corruption for whatever reason, user can delete build dir to start fresh.

+++ mostlyportable-gcc.sh
@@ -211,8 +211,12 @@ _nowhere="$PWD"
       chmod a+x cloog-"${_cloog}".tar.* && tar -xvf cloog-"${_cloog}".tar.* >/dev/null 2>&1
 
       if [[ "$_binutils" = 2.33* ]] && [ "$_valve_patches" == "true" ]; then
-        wget -c -O proton_binutils1.binutilspatch https://raw.githubusercontent.com/ValveSoftware/Proton/3ad34a0b3f41bac60caea39c742de69cb0e50895/mingw-w64-patches/binutils-0001.patch
-        wget -c -O proton_binutils2.binutilspatch https://raw.githubusercontent.com/ValveSoftware/Proton/3ad34a0b3f41bac60caea39c742de69cb0e50895/mingw-w64-patches/binutils-0002.patch
+        if [ ! -e "${_nowhere}/build/proton_binutils1.binutilspatch" ]; then
+          wget -c -O proton_binutils1.binutilspatch https://raw.githubusercontent.com/ValveSoftware/Proton/3ad34a0b3f41bac60caea39c742de69cb0e50895/mingw-w64-patches/binutils-0001.patch
+        fi
+        if [ ! -e "${_nowhere}/build/proton_binutils2.binutilspatch" ]; then
+          wget -c -O proton_binutils2.binutilspatch https://raw.githubusercontent.com/ValveSoftware/Proton/3ad34a0b3f41bac60caea39c742de69cb0e50895/mingw-w64-patches/binutils-0002.patch
+        fi
       fi
       _path_hack="${_path_hack_prefix}${_dstdir}/i686-w64-mingw32:${_dstdir}/x86_64-w64-mingw32:${_dstdir}/libexec:${_dstdir}/bin:${_dstdir}/lib:${_dstdir}/include:${PATH}"
     else
